Add secure vault support for proxy services Fixes: https://github.com/wso2/product-ei/issues/4093
